backend/requirements.txt
# Core backend
fastapi
uvicorn            # runs the FastAPI development server
pydantic
python-dotenv      # loads .env file for API keys and config

# Document generation
python-docx        # programmatic DOCX construction (fonts, spacing, layout in Python)
weasyprint         # HTML → PDF rendering

# AI
openai             # works with both OpenAI and Ollama (same SDK, different base URL)

# API / server
python-multipart   # required by FastAPI to parse form data
slowapi            # rate limiting for the /generate endpoint

# Testing
pytest             # test runner — finds and runs all test_*.py files automatically
httpx              # required by FastAPI's TestClient for making in-process HTTP requests